+++ src/lib/Scenes/Artwork/Components/AuctionCountDownTimer.tsx
@@ -1,54 +1,18 @@
 import { TimeRemaining } from "@artsy/palette"
 import { AuctionCountDownTimer_artwork } from "__generated__/AuctionCountDownTimer_artwork.graphql"
-import { DateTime } from "luxon"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
+import { AuctionState } from "./CommercialInformation"
 
 interface AuctionCountDownTimerProps {
   artwork: AuctionCountDownTimer_artwork
+  auctionState: AuctionState
 }
 
-interface AuctionCountDownTimerState {
-  auctionStatus: null | "hasStarted" | "hasEnded" | "hasNotStarted"
-}
-
-export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimerProps, AuctionCountDownTimerState> {
-  state = {
-    auctionStatus: null,
-  }
-
-  interval = null
-
-  componentDidMount() {
-    const { sale } = this.props.artwork
-    if (!sale || !sale.startAt || !sale.endAt) {
-      return
-    }
-
-    this.interval = setInterval(() => {
-      if (DateTime.local() > DateTime.fromISO(sale.startAt)) {
-        this.setState({
-          auctionStatus: "hasStarted",
-        })
-      } else if (DateTime.local() > DateTime.fromISO(sale.endAt)) {
-        this.setState({
-          auctionStatus: "hasEnded",
-        })
-      } else {
-        this.setState({
-          auctionStatus: "hasNotStarted",
-        })
-      }
-    }, 1000)
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.interval)
-  }
-
+export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimerProps> {
   render() {
+    const { auctionState } = this.props
     const { sale } = this.props.artwork
-    const { auctionStatus } = this.state
 
     if (!sale || !sale.startAt || !sale.endAt) {
       return null
@@ -60,7 +24,7 @@ export class AuctionCountDownTimer extends React.Component<AuctionCountDownTimer
         labelWithoutTimeRemaining={sale.formattedStartDateTime}
         labelFontSize="2"
         timerFontSize="4t"
-        countdownEnd={auctionStatus === "hasNotStarted" ? sale.startAt : sale.endAt}
+        countdownEnd={auctionState === "isPreview" ? sale.startAt : sale.endAt}
         highlight="black100"
       />
     )