+++ src/lib/Scenes/Artwork/Components/CommercialInformation.tsx
@@ -1,6 +1,7 @@
 import { Box, Sans, Spacer } from "@artsy/palette"
 import { CommercialInformation_artwork } from "__generated__/CommercialInformation_artwork.graphql"
 import { capitalize } from "lodash"
+import { DateTime } from "luxon"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { ArtworkExtraLinksFragmentContainer as ArtworkExtraLinks } from "./ArtworkExtraLinks"
@@ -14,19 +15,42 @@ interface CommercialInformationProps {
   artwork: CommercialInformation_artwork
 }
 
+export type AuctionState = "hasStarted" | "isLive" | "hasEnded" | "isPreview" | null
+
 interface CommercialInformationState {
   editionSetID: string
+  auctionState: AuctionState
+}
+
+const initialAuctionState = sale => {
+  if (sale && sale.isAuction) {
+    if (sale.isClosed) {
+      return "hasEnded"
+    } else if (sale.isLiveOpen) {
+      return "isLive"
+    } else if (sale.isPreview) {
+      return "isPreview"
+    } else if (DateTime.local() > DateTime.fromISO(sale.startAt)) {
+      return "hasStarted"
+    }
+  } else {
+    return null
+  }
 }
 
 export class CommercialInformation extends React.Component<CommercialInformationProps, CommercialInformationState> {
   state = {
     editionSetID: null,
+    auctionState: initialAuctionState(this.props.artwork.sale) as AuctionState,
   }
+
+  interval = null
 
   renderSingleEditionArtwork = () => {
     const { artwork } = this.props
+    const { auctionState } = this.state
     let saleMessage
-    if (artwork.isInAuction && artwork.sale && artwork.sale.isClosed) {
+    if (artwork.isInAuction && artwork.sale && auctionState === "hasEnded") {
       saleMessage = "Bidding closed"
     } else if (artwork.saleMessage === "Contact For Price") {
       saleMessage = "Contact for price"
@@ -46,10 +70,11 @@ export class CommercialInformation extends React.Component<CommercialInformation
 
   renderPriceInformation = () => {
     const { artwork } = this.props
+    const { auctionState } = this.state
     const { isInAuction, isForSale } = artwork
 
-    if (isInAuction && isForSale) {
-      return <AuctionPrice artwork={artwork} />
+    if (isInAuction && isForSale && auctionState !== "isLive") {
+      return <AuctionPrice artwork={artwork} auctionState={auctionState as AuctionState} />
     } else if (artwork.editionSets && artwork.editionSets.length > 1) {
       return this.renderEditionSetArtwork()
     } else {
@@ -71,12 +96,58 @@ export class CommercialInformation extends React.Component<CommercialInformation
     )
   }
 
+  componentDidMount = () => {
+    const { artwork } = this.props
+    const { sale } = artwork
+
+    if (sale && sale.isAuction) {
+      this.setState({
+        auctionState: initialAuctionState(sale),
+      })
+
+      if (sale.isClosed) {
+        return
+      }
+
+      // Set intervals to check for state changes
+      if (sale.isPreview && sale.startAt) {
+        this.interval = setInterval(() => {
+          if (DateTime.local() > DateTime.fromISO(sale.startAt)) {
+            this.setState({
+              auctionState: "hasStarted",
+            })
+          }
+        }, 1000)
+      } else if (sale.liveStartAt) {
+        this.interval = setInterval(() => {
+          if (DateTime.local() > DateTime.fromISO(sale.liveStartAt)) {
+            this.setState({
+              auctionState: "isLive",
+            })
+          }
+        }, 1000)
+      } else if (sale.endAt) {
+        this.interval = setInterval(() => {
+          if (DateTime.local() > DateTime.fromISO(sale.endAt)) {
+            this.setState({
+              auctionState: "hasEnded",
+            })
+          }
+        }, 1000)
+      }
+    }
+  }
+
+  componentWillUnmount = () => {
+    clearInterval(this.interval)
+  }
+
   render() {
     const { artwork } = this.props
-    const { editionSetID } = this.state
+    const { editionSetID, auctionState } = this.state
     const { isAcquireable, isOfferable, isInquireable, isInAuction, sale, isForSale } = artwork
 
-    const isBiddableInAuction = isInAuction && sale && !sale.isClosed && isForSale
+    const isBiddableInAuction = isInAuction && sale && auctionState !== "hasEnded" && isForSale
     const canTakeCommercialAction = isAcquireable || isOfferable || isInquireable || isBiddableInAuction
     const artistIsConsignable = artwork.artists.filter(artist => artist.isConsignable).length
 
@@ -87,19 +158,19 @@ export class CommercialInformation extends React.Component<CommercialInformation
           {canTakeCommercialAction && (
             <>
               <Spacer mb={2} />
-              <CommercialButtons artwork={artwork} editionSetID={editionSetID} />
+              <CommercialButtons artwork={artwork} auctionState={auctionState} editionSetID={editionSetID} />
             </>
           )}
           {isBiddableInAuction && (
             <>
               <Spacer mb={2} />
-              <AuctionCountDownTimer artwork={artwork} />
+              <AuctionCountDownTimer artwork={artwork} auctionState={auctionState} />
             </>
           )}
           {(!!artistIsConsignable || isAcquireable || isOfferable || isBiddableInAuction) && (
             <>
               <Spacer mb={2} />
-              <ArtworkExtraLinks artwork={artwork} />
+              <ArtworkExtraLinks artwork={artwork} auctionState={auctionState} />
             </>
           )}
         </Box>
@@ -129,6 +200,12 @@ export const CommercialInformationFragmentContainer = createFragmentContainer(Co
 
       sale {
         isClosed
+        isAuction
+        isLiveOpen
+        isPreview
+        liveStartAt
+        endAt
+        startAt
       }
 
       ...CommercialButtons_artwork