+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 - Remove mobx - ds300
 - Add missing auction analytics events - ashkan18
+- Adds Artwork Auctions refresh on state change - kierangillen
 
 ### 1.17.3